+++ AssetsFlow.Client.React/src/Components/Journal/DataElement/DataElement.jsx
@@ -5,9 +5,11 @@ import { ProcessingStatus } from "@constants/Constants";
 import ProcessingAndSuccessMessage from "@components/Processing/ProcessingAndSuccessMessage";
 import ValueWrapper from "./ValueWrapper";
 import { useContentUpdateMutation } from "@hooks/Journal/Entry/useContentUpdateMutation";
+import { dataElementContentParser } from "@services/dataElementContentParser";
 
-function DataElement({ cellInfo, onCellUpdate }) {
-  const isOverview = onCellUpdate === undefined;
+function DataElement({ cellInfo, onCellUpdate, overviewType }) {
+  const isOverview = overviewType != Constants.OverviewType.NONE
+  const { contentValue } = dataElementContentParser(cellInfo);
 
   const { contentUpdateMutation, processingStatus } = useContentUpdateMutation(
     cellInfo,
@@ -22,27 +24,30 @@ function DataElement({ cellInfo, onCellUpdate }) {
     contentUpdateMutation.mutate(newValue);
   };
 
-  const titleText = isOverview
-    ? `${cellInfo[Constants.DATA_TITLE_STRING]}:`
-    : cellInfo[Constants.DATA_TITLE_STRING];
-
   const containerStyle = {
-    pointerEvents: processingStatus === ProcessingStatus.PROCESSING ? "none" : "auto",
+    pointerEvents:
+      processingStatus === ProcessingStatus.PROCESSING ? "none" : "auto",
     display: "flex",
     gap: "5px",
     flexDirection: isOverview ? "row" : "column",
-    maxWidth: isOverview ? 350 : 150,
-    minWidth : 75,
+    maxWidth: isOverview ? 250 : 150,
+    minWidth: 75,
+    border: !contentValue && overviewType != Constants.OverviewType.TRADE_OVERVIEW ? "1px solid red" : "none",
   };
 
   return (
-    <Paper shadow="xs" px={5} radius="md" withBorder
+    <Paper
+      shadow="xs"
+      px={5}
+      radius="md"
+      withBorder
       className="container-with-centered-content"
       style={containerStyle}
     >
       <Text className="no-overflow-text-style">
-        {titleText}
+        {cellInfo[Constants.DATA_TITLE_STRING]} {isOverview ? ":" : ""}
       </Text>
+
       <ValueWrapper
         cellInfo={cellInfo}
         onValueChangeInitiated={!isOverview ? initiateMutation : undefined}