+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Changelog
 
+## [0.9.12] - 2025-05-26
+### Added
+- Red border on missing data, when you open a trade
+
+## [0.9.11] - 2025-06-05
+### Fixed
+no restriction on width for collapsed element
+
 ## [0.9.1] - 2025-05-26
 ### Added
 - Versions info on the footer